requirements.txt
tensorflow-cpu==2.18.0
numpy
pandas
streamlit
boto3
requests
